refactor: moved Settings page and ProfileImageUpload component into unique module folder
